:traffic_light: Fix routings
