+++ app/routers/serp.py
@@ -7,7 +7,7 @@ from fastapi import APIRouter
 router = APIRouter()
 
 
-@router.get("/serp/{task_id}", response_model=List[SearchResult])
+@router.get("/tasks/{task_id}/serp", response_model=List[SearchResult])
 def get_serp(task_id: int, offset: int = 0):
     repo = SerpRepositoryImpl()
     su = SerpUsecase(repo)
